Moved camera viewmodel to proper file
