+++ app/src/main/java/com/snaptrash/snaptrash/viewmodel/CameraViewModel.kt
@@ -1,0 +1,16 @@
+package com.snaptrash.snaptrash.viewmodel
+
+import android.net.Uri
+import android.util.Log
+import androidx.compose.runtime.mutableStateOf
+import androidx.lifecycle.ViewModel
+
+class CameraViewModel: ViewModel(){
+    var photoUri = mutableStateOf(Uri.EMPTY)
+    var takeImage = mutableStateOf(true)
+    fun handleImageCapture(uri: Uri) {
+        Log.i("kilo", "Image captured: $uri")
+        photoUri.value = uri
+        takeImage.value = false
+    }
+}
